rename integration_tests to integrationtests to match Go naming conventions
