+++ integrationtests/setup.sh
@@ -4,8 +4,8 @@ set -e
 echo "Running setup script..."
 
 CURRENT_DIR=$(pwd)
-if [ $(basename "$CURRENT_DIR") != "integration_tests" ]; then
-    echo "Please run this script from the integration_tests directory."
+if [ $(basename "$CURRENT_DIR") != "integrationtests" ]; then
+    echo "Please run this script from the integrationtests directory."
     exit 1
 fi
